djangoapp/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5
psycopg2-binary==2.9.9
pybase64==1.3.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
